+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Common and useful classes, methods, exceptions etc.
 
+# 0.1.1
+
+1. TravisCI > run using PHP 7.2 too
+2. ValueObject > class Version > represents version of software
+3. Move version of this package to `VERSION` file (from `composer.json` file)
+
 # 0.1.0
 
 1. Composer > support/require PHP 5.6+ (instead of 5.5.9+)